airflow/requirements.txt
# Airflow container requirements
# Installed on top of apache/airflow:2.8.1-python3.11 base image

# GCP integrations
google-cloud-storage==2.16.0
google-cloud-bigquery==3.20.1
google-cloud-bigquery-storage==2.25.0
pyarrow==15.0.2
db-dtypes==1.2.0

# HTTP
requests==2.31.0

# Data processing
pandas==2.2.2

# dbt (BigQuery adapter)
dbt-core==1.7.14
dbt-bigquery==1.7.9
